Restructure project: delete examples directory, test directory is now self-reliant
