+++ .gitignore
@@ -2,10 +2,6 @@
 # NodeJS
 #
 node_modules
-examples/jsonToWar/output/war3*
-examples/warToJson/files/output/*.json
-examples/warToJson/files/output/object/*.json
-examples/warToJson/files/output/other/*.json
 
 #
 # Build, testing, and coverage
