--- examples/package.json
@@ -1,17 +1,0 @@
-{
-  "name": "examples",
-  "version": "2.0.0",
-  "description": "Examples of using WC3MapTranslator to generate translate between war3map and JSON",
-  "main": "index.js",
-  "scripts": {
-    "build": "tsc --build",
-    "start": "npm run build && node index.js",
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "author": "ChiefOfGxBxL",
-  "license": "MIT",
-  "dependencies": {
-    "@types/node": "^12.12.21",
-    "typescript": "^3.7.4"
-  }
-}
