--- examples/tsconfig.json
@@ -1,27 +1,0 @@
-{
-    "compilerOptions": {
-        "types": [
-            "mocha",
-            "node"
-        ],
-        "traceResolution": true,
-        "declaration": true,
-        "declarationMap": true,
-        "target": "es2016",
-        "module": "commonjs",
-        "outDir": "./",
-        "rootDir": "./",
-        "sourceMap": true,
-        "incremental": true,
-        "tsBuildInfoFile": "./.tsbuildinfo",
-        "esModuleInterop": true,
-        "noEmit": false,
-    },
-    "references": [
-        { "path": "../" }
-    ],
-    "exclude": [],
-    "include": [
-        "./**/*"
-    ]
-}
