+++ test/data/cameras.json
