+++ test/data/object-abilities.json
