+++ test/data/war3map.doo
