+++ test/data/war3map.shd
