+++ test/data/war3map.w3h
