+++ test/data/war3map.w3q
