+++ test/data/war3map.w3t
