+++ test/data/war3map.wts
